* commit '555018d0440e342199b68eeb59c532b66999f642':
  Fix IO_INPUT with 64-bit.

+++ rs/java/android/renderscript/Allocation.java
@@ -1868,7 +1868,7 @@ public class Allocation extends BaseObj {
         }
     }
 
-    static void sendBufferNotification(int id) {
+    static void sendBufferNotification(long id) {
         synchronized(mAllocationMap) {
             Allocation a = mAllocationMap.get(new Long(id));
